mobot/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e git+https://github.com/mobilecoinofficial/full-service.git@7616af6e6f9f37d9b922836215f16e724d110ca7#egg=mobilecoin&subdirectory=cli
amqp==5.0.6; python_version >= '3.6'
appnope==0.1.2; sys_platform == 'darwin'
asgiref==3.4.0; python_version >= '3.6'
attrs==21.2.0
azt==0.0.7; python_version >= '3'
babel==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cython==0.29.23; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
dacite==1.6.0
dataclasses-json==0.5.4
decorator==5.0.9; python_version >= '3.5'
django-address==0.2.5
django-celery-results==2.1.0
django-environ==0.4.5
django-extensions==3.1.3
django-fsm==2.7.1
django-money[exchange]==2.0
django-phonenumber-field==5.2.0
django==3.2.4
djangorestframework-dataclasses==0.9
djangorestframework==3.12.4; python_version >= '3.5'
expiringdict==1.2.1
google-i18n-address==2.4.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipython-genutils==0.2.0
ipython==7.25.0
jedi==0.18.0; python_version >= '3.6'
kombu==5.1.0; python_version >= '3.6'
marshmallow-enum==1.5.1
marshmallow==3.12.1; python_version >= '3.5'
matplotlib-inline==0.1.2; python_version >= '3.5'
mypy-extensions==0.4.3
parso==0.8.2; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
phonenumbers==8.12.26
pickleshare==0.7.5
prompt-toolkit==3.0.19; python_full_version >= '3.6.1'
psycopg2==2.9.1
ptyprocess==0.7.0
py-moneyed==1.2
pyfiglet==0.8.post1
pygments==2.9.0; python_version >= '3.5'
python-dotenv==0.18.0
python-json-logger==2.0.1
pytz==2021.1
pyzt==1.0.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
stringcase==1.2.0
termcolor==1.1.0
traitlets==5.0.5; python_version >= '3.7'
typedate==1.1.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uwsgi==2.0.19.1
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
